requirements.txt
fastapi>=0.115
uvicorn>=0.34
sqlalchemy>=2.0
httpx>=0.27
python-dotenv>=1.0
jinja2>=3.1
python-multipart>=0.0.9
twilio>=9.0.0
pandas>=2.2.0
openpyxl>=3.1.0